requirements.txt
hyperliquid-python-sdk
websockets>=12.0
python-dotenv>=1.0